Refactor metap to avoid passing doc everywhere
This not only makes the code simpler, but also avoids the need to
maintain weak hashmaps (which was, frankly speaking, insane).

The only place where getting rid of explicit doc passing was annoying
is prettyprinting of ranges, however I figured that out by modularizing
range printing into smaller components.
